feat(functions): added shapefile processing (idas-221)
- refactored malware-detected code to modularize it
- added readme for future enhancements
