+++ apps/functions/applications-background-jobs/malware-detected/index.js
@@ -2,65 +2,7 @@ import { sendDocumentStateAction } from './src/back-office-api-client.js';
 import { isScannedFileHtml, isUploadedHtmlValid } from '../common/html-validation.js';
 import { getPublishedStatusFromScanResult } from './src/util.js';
 import { guidFromBlobURI } from '../common/util.js';
-import { enqueueShapefileForProcessing } from './src/shapefile-queue.js';
-import {
-	GIS_SHAPEFILE_DOCUMENT_TYPE,
-	DocumentPublishedStatus,
-	GEOJSON_ENQUEUE_REASON
-} from '../common/constants.js';
-
-const ZIP_FILE_REGEX = /\.zip$/i;
-const GEOJSON_FILE_REGEX = /\.geojson($|\?)/i;
-
-/**
- * Pure predicate: returns true when a document should be enqueued for shapefile processing.
- * Extracted from the main handler to make the routing logic independently readable and testable.
- *
- * Conditions (all must hold):
- * - Virus scan passed (publishedStatus === NOT_CHECKED)
- * - Document is typed as a GIS shapefile
- * - documentURI is present (needed by the processing function)
- * - Original filename is a ZIP (not a generated .geojson)
- * - Current blob event is not for a .geojson (prevents re-enqueue feedback loops)
- *
- * @param {string} publishedStatus
- * @param {{ documentType: string|null, documentURI: string|null, originalFilename: string|null } | null | undefined} documentInfo
- * @param {string} blobUri - URI of the blob that triggered the AV scan event
- * @returns {{ enqueue: boolean, reason: string | null }}
- */
-const shouldEnqueueShapefile = (publishedStatus, documentInfo, blobUri) => {
-	const resolvedDocumentUri = documentInfo?.documentURI ?? blobUri;
-
-	const checks = [
-		{
-			invalid: publishedStatus !== DocumentPublishedStatus.NOT_CHECKED,
-			reason: GEOJSON_ENQUEUE_REASON.SCAN_NOT_CLEAN
-		},
-		{
-			invalid: documentInfo?.documentType !== GIS_SHAPEFILE_DOCUMENT_TYPE,
-			reason: GEOJSON_ENQUEUE_REASON.NOT_GIS_SHAPEFILE
-		},
-		{ invalid: !resolvedDocumentUri, reason: GEOJSON_ENQUEUE_REASON.NO_DOCUMENT_URI },
-		{
-			invalid:
-				Boolean(documentInfo?.originalFilename) &&
-				!ZIP_FILE_REGEX.test(documentInfo?.originalFilename ?? ''),
-			reason: GEOJSON_ENQUEUE_REASON.SOURCE_NOT_ZIP
-		},
-		{
-			invalid:
-				GEOJSON_FILE_REGEX.test(blobUri) || GEOJSON_FILE_REGEX.test(resolvedDocumentUri ?? ''),
-			reason: GEOJSON_ENQUEUE_REASON.GEOJSON_EVENT_LOOP_GUARD
-		}
-	];
-
-	const failedCheck = checks.find((check) => check.invalid);
-
-	return {
-		enqueue: !failedCheck,
-		reason: failedCheck ? failedCheck.reason : null
-	};
-};
+import { routeShapefileForProcessing } from './src/shapefile-routing.js';
 
 /**
  * Azure Function triggered by EventGrid antivirus scan completion events.
@@ -104,45 +46,14 @@ export const index = async (context, eventGridEvent) => {
 
 	// Update document status in the DB; the enriched response tells us what kind of document it is
 	const documentInfo = await sendDocumentStateAction(guid, publishedStatus, context.log);
-	const resolvedDocumentUri = documentInfo?.documentURI ?? blobUri;
-	const isSourceZip =
-		!documentInfo?.originalFilename || ZIP_FILE_REGEX.test(documentInfo?.originalFilename);
-	const isGeoJsonEvent = GEOJSON_FILE_REGEX.test(blobUri);
-	const isGeoJsonDocumentUri = GEOJSON_FILE_REGEX.test(resolvedDocumentUri ?? '');
-
-	const { enqueue, reason } = shouldEnqueueShapefile(publishedStatus, documentInfo, blobUri);
-
-	context.log('[SHAPEFILE] Routing decision', {
+	await routeShapefileForProcessing({
 		eventId,
 		guid,
 		publishedStatus,
-		documentType: documentInfo?.documentType ?? null,
-		hasDocumentURI: Boolean(documentInfo?.documentURI),
-		hasOriginalFilename: Boolean(documentInfo?.originalFilename),
-		isSourceZip,
-		isGeoJsonEvent,
-		isGeoJsonDocumentUri,
-		enqueue,
-		reason
+		documentInfo,
+		blobUri,
+		log: context.log
 	});
-
-	if (enqueue) {
-		// shouldEnqueueShapefile guarantees documentInfo is non-null and documentURI is present when enqueue=true
-		const safeDocumentInfo = /** @type {NonNullable<typeof documentInfo>} */ (documentInfo);
-
-		await enqueueShapefileForProcessing({
-			documentId: guid,
-			caseId: safeDocumentInfo.caseId,
-			caseRef: safeDocumentInfo.caseRef,
-			documentURI: safeDocumentInfo.documentURI ?? blobUri,
-			originalFilename: safeDocumentInfo.originalFilename,
-			dateCreated: safeDocumentInfo.dateCreated
-		});
-
-		context.log(`[SHAPEFILE] Enqueued document ${guid} for shapefile processing`);
-	} else {
-		context.log(`[SHAPEFILE] Skipping shapefile queue routing: ${reason} (guid: ${guid})`);
-	}
 
 	context.log('Anti Virus Scan event completed: ', { eventId });
 };
