+++ apps/functions/applications-background-jobs/malware-detected/src/shapefile-routing.js
@@ -1,0 +1,112 @@
+import { enqueueShapefileForProcessing } from './shapefile-queue.js';
+import {
+	GIS_SHAPEFILE_DOCUMENT_TYPE,
+	DocumentPublishedStatus,
+	GEOJSON_ENQUEUE_REASON,
+	SHAPEFILE_SOURCE_ZIP_FILE_REGEX,
+	GEOJSON_FILE_REGEX
+} from '../../common/constants.js';
+
+/**
+ * Determines whether a scanned document should be routed to shapefile processing.
+ *
+ * @param {string} publishedStatus
+ * @param {{ documentType: string|null, documentURI: string|null, originalFilename: string|null } | null | undefined} documentInfo
+ * @param {string} blobUri
+ * @returns {{ enqueue: boolean, reason: string | null }}
+ */
+export const shouldEnqueueShapefile = (publishedStatus, documentInfo, blobUri) => {
+	const resolvedDocumentUri = documentInfo?.documentURI ?? blobUri;
+
+	const checks = [
+		{
+			invalid: publishedStatus !== DocumentPublishedStatus.NOT_CHECKED,
+			reason: GEOJSON_ENQUEUE_REASON.SCAN_NOT_CLEAN
+		},
+		{
+			invalid: documentInfo?.documentType !== GIS_SHAPEFILE_DOCUMENT_TYPE,
+			reason: GEOJSON_ENQUEUE_REASON.NOT_GIS_SHAPEFILE
+		},
+		{ invalid: !resolvedDocumentUri, reason: GEOJSON_ENQUEUE_REASON.NO_DOCUMENT_URI },
+		{
+			invalid:
+				Boolean(documentInfo?.originalFilename) &&
+				!SHAPEFILE_SOURCE_ZIP_FILE_REGEX.test(documentInfo?.originalFilename ?? ''),
+			reason: GEOJSON_ENQUEUE_REASON.SOURCE_NOT_ZIP
+		},
+		{
+			invalid:
+				GEOJSON_FILE_REGEX.test(blobUri) || GEOJSON_FILE_REGEX.test(resolvedDocumentUri ?? ''),
+			reason: GEOJSON_ENQUEUE_REASON.GEOJSON_EVENT_LOOP_GUARD
+		}
+	];
+
+	const failedCheck = checks.find((check) => check.invalid);
+
+	return {
+		enqueue: !failedCheck,
+		reason: failedCheck ? failedCheck.reason : null
+	};
+};
+
+/**
+ * Handles shapefile queue routing for clean GIS shapefiles.
+ *
+ * @param {{
+ *   eventId: string,
+ *   guid: string,
+ *   publishedStatus: string,
+ *   documentInfo: any,
+ *   blobUri: string,
+ *   log: any
+ * }} params
+ */
+export const routeShapefileForProcessing = async ({
+	eventId,
+	guid,
+	publishedStatus,
+	documentInfo,
+	blobUri,
+	log
+}) => {
+	const resolvedDocumentUri = documentInfo?.documentURI ?? blobUri;
+	const isSourceZip =
+		!documentInfo?.originalFilename ||
+		SHAPEFILE_SOURCE_ZIP_FILE_REGEX.test(documentInfo?.originalFilename);
+	const isGeoJsonEvent = GEOJSON_FILE_REGEX.test(blobUri);
+	const isGeoJsonDocumentUri = GEOJSON_FILE_REGEX.test(resolvedDocumentUri ?? '');
+
+	const { enqueue, reason } = shouldEnqueueShapefile(publishedStatus, documentInfo, blobUri);
+
+	log('[SHAPEFILE] Routing decision', {
+		eventId,
+		guid,
+		publishedStatus,
+		documentType: documentInfo?.documentType ?? null,
+		hasDocumentURI: Boolean(documentInfo?.documentURI),
+		hasOriginalFilename: Boolean(documentInfo?.originalFilename),
+		isSourceZip,
+		isGeoJsonEvent,
+		isGeoJsonDocumentUri,
+		enqueue,
+		reason
+	});
+
+	if (!enqueue) {
+		log(`[SHAPEFILE] Skipping shapefile queue routing: ${reason} (guid: ${guid})`);
+		return;
+	}
+
+	const safeDocumentInfo = /** @type {NonNullable<typeof documentInfo>} */ (documentInfo);
+
+	await enqueueShapefileForProcessing({
+		documentId: guid,
+		caseId: safeDocumentInfo.caseId,
+		caseRef: safeDocumentInfo.caseRef,
+		documentURI: safeDocumentInfo.documentURI ?? blobUri,
+		originalFilename: safeDocumentInfo.originalFilename,
+		dateCreated: safeDocumentInfo.dateCreated
+	});
+
+	log(`[SHAPEFILE] Enqueued document ${guid} for shapefile processing`);
+};
